+++ app/jobs/inbound_webhooks/brevo/process_mail_delivery_status_job.rb
@@ -3,8 +3,8 @@ module InboundWebhooks
     class ProcessMailDeliveryStatusJob < ProcessDeliveryStatusJobBase
       private
 
-      def process_invitation
-        Invitations::AssignMailDeliveryStatusAndDate.call(webhook_params: @webhook_params,
+      def assign_delivery_status_and_date
+        InboundWebhooks::Brevo::AssignMailDeliveryStatusAndDate.call(webhook_params: @webhook_params,
                                                           invitation: @invitation)
       end
     end
