+++ db/migrate/20240531140628_add_delivery_status_and_delivery_date_time_to_invitations.rb
@@ -1,6 +1,6 @@
 class AddDeliveryStatusAndDeliveryDateTimeToInvitations < ActiveRecord::Migration[7.1]
   def change
     add_column :invitations, :delivery_status, :string
-    add_column :invitations, :delivery_status_received_at, :datetime
+    add_column :invitations, :delivered_at, :datetime
   end
 end
