+++ spec/services/inbound_webhooks/brevo/assign_sms_delivery_status_and_date_spec.rb
@@ -1,4 +1,4 @@
-describe Invitations::AssignSmsDeliveryStatusAndDate do
+describe InboundWebhooks::Brevo::AssignSmsDeliveryStatusAndDate do
   subject { described_class.call(webhook_params: webhook_params, invitation: invitation) }
 
   let(:webhook_params) { { to: "0601010101", msg_status: "delivered", date: "2023-06-07T12:34:56Z" } }
@@ -8,11 +8,11 @@ describe Invitations::AssignSmsDeliveryStatusAndDate do
   context "when the invitation has a final delivery status" do
     Invitation::FINAL_DELIVERY_STATUS.each do |status|
       it "does not update the invitation if delivery status is #{status}" do
-        invitation.update(delivery_status: status, delivery_status_received_at: Time.zone.parse("2023-06-07T12:00:00Z"))
+        invitation.update(delivery_status: status, delivered_at: Time.zone.parse("2023-06-07T12:00:00Z"))
         subject
         invitation.reload
         expect(invitation.delivery_status).to eq(status)
-        expect(invitation.delivery_status_received_at).to eq(Time.zone.parse("2023-06-07T12:00:00Z"))
+        expect(invitation.delivered_at).to eq(Time.zone.parse("2023-06-07T12:00:00Z"))
       end
     end
   end
@@ -21,14 +21,14 @@ describe Invitations::AssignSmsDeliveryStatusAndDate do
     let(:old_date) { "2022-06-07T12:34:56Z" }
 
     before do
-      invitation.update(delivery_status_received_at: Time.zone.parse("2023-06-08T12:34:56Z"))
+      invitation.update(delivered_at: Time.zone.parse("2023-06-08T12:34:56Z"))
     end
 
     it "does not update the invitation" do
       subject
       invitation.reload
       expect(invitation.delivery_status).to be_nil
-      expect(invitation.delivery_status_received_at).not_to eq(old_date)
+      expect(invitation.delivered_at).not_to eq(old_date)
     end
   end
 
@@ -41,7 +41,7 @@ describe Invitations::AssignSmsDeliveryStatusAndDate do
       )
       subject
       expect(invitation.delivery_status).to be_nil
-      expect(invitation.delivery_status_received_at).to be_nil
+      expect(invitation.delivered_at).to be_nil
     end
   end
 
@@ -49,6 +49,6 @@ describe Invitations::AssignSmsDeliveryStatusAndDate do
     subject
     invitation.reload
     expect(invitation.delivery_status).to eq("delivered")
-    expect(invitation.delivery_status_received_at).to eq(Time.zone.parse("2023-06-07T12:34:56Z"))
+    expect(invitation.delivered_at).to eq(Time.zone.parse("2023-06-07T12:34:56Z"))
   end
 end
